backend/requirements.txt
Django==5.1.4
psycopg2-binary==2.9.9
gunicorn==23.0.0
whitenoise==6.8.2
django-htmx==1.21.0
Markdown==3.7
python-decouple==3.8
Pillow==11.0.0
